Add Interests

+++ repository/article.ts
@@ -5,6 +5,8 @@ import Interest from "../models/interest";
 import Journal from "../models/journal";
 import fs from "fs";
 import path from "path";
+import { IncludeOptions, Op } from "sequelize";
+import ArticleInterest from "../models/article_interest";
 
 import { requestPagination } from "../helper/requestPagination";
 
@@ -13,22 +15,51 @@ class ArticleRepo {
   Author: typeof Author;
   Interest: typeof Interest;
   Journal: typeof Journal;
+  ArticleInterest: typeof ArticleInterest;
   constructor() {
     this.Article = Article;
     this.Author = Author;
     this.Interest = Interest;
     this.Journal = Journal;
+    this.ArticleInterest = ArticleInterest;
   }
 
   allArticles = async (filters: {}) => {
     try {
       const { limit, sort, page, sortBy } = requestPagination(filters);
 
+      let where = filters["search"]
+        ? {
+            [Op.or]: {
+              title: { [Op.like]: `%${filters["search"]}%` },
+              abstract: { [Op.like]: `%${filters["search"]}%` },
+              topic: { [Op.like]: `%${filters["search"]}%` },
+            },
+          }
+        : {};
+
       let articles = await db.transaction(async (transaction) => {
         return await this.Article.findAndCountAll({
+          where: where,
           offset: (page - 1) * limit,
           limit,
-          include: [this.Journal],
+          include: [
+            {
+              model: this.Journal,
+              as: "journal",
+              transaction,
+            } as IncludeOptions,
+            {
+              model: this.Author,
+              as: "authors",
+              transaction,
+            } as IncludeOptions,
+            {
+              model: this.Interest,
+              as: "interests",
+              transaction,
+            } as IncludeOptions,
+          ],
           order: [[sortBy, sort]],
           distinct: true,
           transaction,
@@ -48,7 +79,23 @@ class ArticleRepo {
           where: {
             id: id,
           },
-          include: [this.Journal],
+          include: [
+            {
+              model: this.Journal,
+              as: "journal",
+              transaction,
+            } as IncludeOptions,
+            {
+              model: this.Author,
+              as: "authors",
+              transaction,
+            } as IncludeOptions,
+            {
+              model: this.Interest,
+              as: "interests",
+              transaction,
+            } as IncludeOptions,
+          ],
           transaction,
         });
       });
@@ -61,13 +108,30 @@ class ArticleRepo {
 
   createArticle = async (articleData: any) => {
     try {
-      let article = await db.transaction(async () => {
+      let article = await db.transaction(async (transaction) => {
         return await this.Article.create(articleData);
       });
 
-      Object.assign(articleData.authors, { article_id: article["id"] });
+      await articleData["authors"].map(async (author: any) => {
+        Object.assign(author, { article_id: article["id"] });
 
-      console.log(articleData.authors);
+        await this.Author.create(author);
+      });
+
+      await articleData["interests"].map(async (interest: any) => {
+        const check = await this.Interest.findOne({
+          where: {
+            name: interest["name"],
+          },
+        });
+
+        if (check) {
+          await this.ArticleInterest.create({
+            article_id: article["id"],
+            interest_id: check["id"],
+          });
+        }
+      });
 
       return article;
     } catch (error: any) {
@@ -81,6 +145,57 @@ class ArticleRepo {
         return await article.update(articleData, transaction);
       });
 
+      if (
+        articleData["interests"] != undefined &&
+        articleData["interests"] != null
+      ) {
+        await articleData["interests"].map(async (interest: any) => {
+          const check = await this.Interest.findOne({
+            where: {
+              name: interest["name"],
+            },
+          });
+
+          if (check) {
+            const isExists = await this.ArticleInterest.findOne({
+              where: {
+                article_id: article["id"],
+                interest_id: check["id"],
+              },
+            });
+
+            if (!isExists) {
+              await this.ArticleInterest.create({
+                article_id: article["id"],
+                interest_id: check["id"],
+              });
+            }
+          }
+        });
+      }
+
+      if (
+        articleData["authors"] != undefined &&
+        articleData["authors"] != null
+      ) {
+        await articleData["authors"].map(async (author: any) => {
+          Object.assign(author, { article_id: article["id"] });
+
+          const authorCheck = await this.Author.findOne({
+            where: {
+              article_id: update["id"],
+              email: author["email"],
+            },
+          });
+
+          if (!authorCheck) {
+            await this.Author.create(author);
+          } else {
+            await authorCheck.update(author);
+          }
+        });
+      }
+
       return update;
     } catch (error: any) {
       return error["message"];
@@ -89,21 +204,31 @@ class ArticleRepo {
 
   deleteArticle = async (article: any) => {
     try {
-      fs.unlink(
+      if (
+        article["file"] != null &&
         article["file"].replace(
           `http://127.0.0.1:5000`,
           path.join(__dirname + "/../static")
-        ),
-        (err) => {
-          if (err) {
-            return;
+        ) != undefined
+      ) {
+        fs.unlink(
+          article["file"].replace(
+            `http://127.0.0.1:5000`,
+            path.join(__dirname + "/../static")
+          ),
+          (err) => {
+            if (err) {
+              return;
+            }
           }
-        }
-      );
+        );
+      }
 
-      return await db.transaction(async () => {
+      let results = await db.transaction(async () => {
         return await article.destroy();
       });
+
+      return results;
     } catch (error: any) {
       return error["message"];
     }
